add js docs, break out relevanceScore unit tests into their own file
