+++ lib/util/proximity.js
@@ -76,10 +76,10 @@ function center2zxy(center, z) {
  * Combine score and distance into a single score for sorting.
  *
  * @param {Number} score The score of the feature
+ * @param {Number} minScore The lowest score of all features in all indexes.
  * @param {Number} maxScore The maximum score of all features in all indexes
  * @param {Number} dist The distance from the feature to the proximity point in miles.
- * @param {String} type The feature type. If the feature has multiple types,
- *  use the last value
+ * @param {String} zoom The vector tile zoom level of the index the feature is part of.
  * @return {Number} proximity adjusted score value between 1 and 121
  */
 function scoredist(score, minScore, maxScore, dist, zoom) {
@@ -94,6 +94,7 @@ function scoredist(score, minScore, maxScore, dist, zoom) {
  * between a given feature's score and the highest score of all features in all indexes.
  *
  * @param {Number} score The score of the feature.
+ * @param {Number} minScore The lowest score of all features in all indexes.
  * @param {Number} maxScore The highest score of all features in all indexes.
  * @return {Number} scaled score value
  */
@@ -110,8 +111,7 @@ function scoreWeight(score, minScore, maxScore) {
  * distance weighting decays rapidly, then levels out towards the edge of the
  * proximity radius.
  * @param {Number} dist The distance in miles from the feature to the proximity point
- * @param {String} type The feature type. If the feature has multiple types,
- *  use the last value
+ * @param {String} zoom The vector tile zoom level of the index the feature is part of.
  * @return {Number} scaled score value
  */
 function distWeight(dist, zoom) {
@@ -123,8 +123,7 @@ function distWeight(dist, zoom) {
 
 /**
  * Set proximity radius by feature type
- * @param {String} type The feature type. If the feature has multiple types,
- *  use the last value
+ * @param {String} zoom The vector tile zoom level of the index the feature is part of.
  * @return {Number} scaled radius in miles
  */
 function scaleRadius(zoom) {
@@ -184,9 +183,21 @@ function distscore(dist, score) {
     return Math.round(score * (1000 / (Math.max(dist, 50))) * 10000) / 10000;
 }
 
+/**
+ * Combine scoredist and relevance into a single score for sorting. Apply penalties.
+ *
+ * @param {Number} relev The score of the feature
+ * @param {Number} scoredist proximity adjusted score value between 1 and 121
+ * @param {Number} address The carmen:address property of the feature. Equal to
+ * the street number if the user's query matches an address number and point in
+ * the address cluster.
+ * @param {Boolean} omitted True if the street's nearest endpoints match.
+ * @param {String} ghost True if the feature's carmen:score property is < 0.
+ * @return {Number} composite scoredist and relevance score between 0 and 1
+ */
 function relevanceScore(relev, scoredist, address, omitted, ghost) {
     // add a slight penalty to features with carmen:address === null
-    // ommitted geometries, or scores of < 0
+    // omitted geometries, or scores of < 0
     if (address === null) relev = Math.max(0, relev - 0.005);
     if (omitted) relev = Math.max(0, relev - 0.01);
     if (ghost) relev = Math.max(0, relev - 0.01);
